+++ lib/repositories/index.dart
@@ -2,6 +2,6 @@ export 'completion_repository.dart';
 export 'conversation_repository.dart';
 export 'handlers/index.dart';
 export 'image_generation_repository.dart';
-export 'listener_repository.dart';
 export 'threads_repository.dart';
+export 'transcriber_repository.dart';
 export 'tts_repository.dart';