+++ lib/usecases/audio/transcribe.dart
@@ -1,15 +1,10 @@
 import 'dart:io';
 
-import '../../data/configuration.dart';
-import '../../data/enums/audio_model.dart';
 import '../../factories/create_listener_repository.dart';
 
+@Deprecated('Use "createListenerRepository" to create a transcribe class')
 Future<String> transcribe(File file) async {
   var rep = createListenerRepository();
-  var text = await rep.listen(
-    audio: file,
-    model: AudioModel.whisper1,
-    language: VitGptConfiguration.transcriptionLanguage,
-  );
+  var text = await rep.transcribeFromFile(file);
   return text;
 }